+++ packages/typespec-go/test/local/armtest/zz_configurations_client_example_test.go
@@ -37,6 +37,29 @@ func ExampleConfigurationsClient_GetStreamingContent() {
 	// }
 }
 
+// Generated from example definition: 2025-01-01/Configurations_getTextContent.json
+func ExampleConfigurationsClient_GetTextContent() {
+	cred, err := azidentity.NewDefaultAzureCredential(nil)
+	if err != nil {
+		log.Fatalf("failed to obtain a credential: %v", err)
+	}
+	ctx := context.Background()
+	clientFactory, err := armtest.NewClientFactory(cred, nil)
+	if err != nil {
+		log.Fatalf("failed to create client: %v", err)
+	}
+	res, err := clientFactory.NewConfigurationsClient("00000000-0000-0000-0000-000000000000").GetTextContent(ctx, "myResourceGroup", "myConfiguration", nil)
+	if err != nil {
+		log.Fatalf("failed to finish the request: %v", err)
+	}
+	// You could use response here. We use blank identifier for just demo purposes.
+	_ = res
+	// If the HTTP response code is 200 as defined in example definition, your response structure would look as follows. Please pay attention that all the values in the output are fake values for just demo purposes.
+	// res = armtest.ConfigurationsClientGetTextContentResponse{
+	// 	Value: to.Ptr("config text"),
+	// }
+}
+
 // Generated from example definition: 2025-01-01/Configurations_putStreamingContent.json
 func ExampleConfigurationsClient_PutStreamingContent() {
 	cred, err := azidentity.NewDefaultAzureCredential(nil)